fpsensor: Use consistent unsigned type for finger index
This is a followup to https://crrev.com/c/3872809, which changed the
variable tracking the number of fingers to a consistent type (uint16_t).

BRANCH=none
BUG=b:244781166
TEST=./test/run_device_tests.py --board dartmonkey
         => PASS
TEST=./test/run_device_tests.py --board bloonchipper
         => PASS
TEST=Flash firmware on gimble, enroll, lock, unlock

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4382560

+++ common/fpsensor/fpsensor.cc
@@ -412,7 +412,7 @@ static enum ec_status fp_command_frame(struct host_cmd_handler_args *args)
 	const auto *params =
 		static_cast<const struct ec_params_fp_frame *>(args->params);
 	void *out = args->response;
-	uint32_t idx = FP_FRAME_GET_BUFFER_INDEX(params->offset);
+	uint16_t idx = FP_FRAME_GET_BUFFER_INDEX(params->offset);
 	uint32_t offset = params->offset & FP_FRAME_OFFSET_MASK;
 	uint32_t size = params->size;
 	uint16_t fgr;
@@ -579,7 +579,7 @@ static enum ec_status fp_command_template(struct host_cmd_handler_args *args)
 	uint32_t size = params->size & ~FP_TEMPLATE_COMMIT;
 	bool xfer_complete = params->size & FP_TEMPLATE_COMMIT;
 	uint32_t offset = params->offset;
-	uint32_t idx = templ_valid;
+	uint16_t idx = templ_valid;
 	uint8_t key[SBP_ENC_KEY_LEN];
 	struct ec_fp_template_encryption_metadata *enc_info;
 	int ret;
